requirements.txt
openai==0.27.2
pinecone-client==2.2.1
datasets==2.11.0
ipykernel==6.22.0
jupyter==1.0.0
gitpython==3.1.31
SQLAlchemy==2.0.7
alembic==1.10.2